+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add `Tsrynge` and `reflect metadata` to implement server container dependencies
   - Enable to implement Repositories / Services classes used on backend
 - Implementation of basic CRUD Value Repository and api endpoints
+- add base context api code for context hook implementations
 
 ## 2021.11.16